mcp/custom-server/pyproject.toml
[project]
name = "custom-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.10.1",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
